@@ -1,9 +1,9 @@
 package in.apps.maitreya.travelalarm;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.view.View;
 
 public class FeedbackActivity extends AppCompatActivity {
